// backend/src/interfaces/student.ts
export interface Student {
    id?: string;
    name?: string;
    pinyin?: string;
    idCard?: string;
    unit?: string;
    dept?: string;
    job?: string;
    address?: string;
    phone?: string;
    email?: string;
    emailCode?: string;
    company?: string;
    checked?: number | string;
    signed?: number | string;
}